Crate beefy_gadget

Modules

communication

Communication streams for the BEEFY networking protocols.

import

justification

Macros

metric_inc

metric_set

Structs

BeefyNetworkParams

BEEFY gadget network parameters.

BeefyParams

BEEFY gadget initialization parameters.

BeefyRPCLinks

Links used by the BEEFY RPC layer, from the BEEFY background voter.

BeefyVoterLinks

Links between the block importer, the background voter and the RPC layer, to be used by the voter.

Traits

Client

A convenience BEEFY client trait that defines all the type bounds a BEEFY client has to satisfy. Ideally that should actually be a trait alias. Unfortunately as of today, Rust does not allow a type alias to be used as a trait bound. Tracking issue is https://github.com/rust-lang/rust/issues/41517.

Functions

beefy_block_import_and_links

Make block importer and link half necessary to tie the background voter to it.

gossip_protocol_name

Name of the votes gossip protocol used by BEEFY.

justifs_protocol_name

Name of the BEEFY justifications request-response protocol.

start_beefy_gadget

Start the BEEFY gadget.