Module beefy_gadget::communication::notification

Structs

BeefyBestBlockTracingKey

Provides tracing key for BEEFY best block stream.

BeefyVersionedFinalityProofTracingKey

Provides tracing key for BEEFY versioned finality proof stream.

Type Definitions

BeefyBestBlockSender

The sending half of the notifications channel(s) used to send notifications about best BEEFY block from the gadget side.

BeefyBestBlockStream

The receiving half of a notifications channel used to receive notifications about best BEEFY blocks determined on the gadget side.

BeefyVersionedFinalityProofSender

The sending half of the notifications channel(s) used to send notifications about versioned finality proof generated at the end of a BEEFY round.

BeefyVersionedFinalityProofStream

The receiving half of a notifications channel used to receive notifications about versioned finality proof generated at the end of a BEEFY round.