Trait beefy_gadget::Client

pub trait Client<B, BE>: BlockchainEvents<B> + HeaderBackend<B> + Finalizer<B, BE> + ProvideRuntimeApi<B> + Send + Sync where
    B: Block,
    BE: Backend<B>, { }

A convenience BEEFY client trait that defines all the type bounds a BEEFY client has to satisfy. Ideally that should actually be a trait alias. Unfortunately as of today, Rust does not allow a type alias to be used as a trait bound. Tracking issue is https://github.com/rust-lang/rust/issues/41517.