Crate beefy_gadget

Modules

notification

Macros

metric_inc

metric_set

Structs

BeefyParams

BEEFY gadget initialization parameters.

Constants

BEEFY_PROTOCOL_NAME

Traits

Client

A convenience BEEFY client trait that defines all the type bounds a BEEFY client has to satisfy. Ideally that should actually be a trait alias. Unfortunately as of today, Rust does not allow a type alias to be used as a trait bound. Tracking issue is https://github.com/rust-lang/rust/issues/41517.

Functions

beefy_peers_set_config

Returns the configuration value to put in sc_network::config::NetworkConfiguration::extra_sets.

start_beefy_gadget

Start the BEEFY gadget.